requirements.txt
Authlib==1.3.0
Faker==25.0.0
Flask==3.0.2
Django==5.0.3
configparser==6.0.1
openai==1.30.1
pandas==2.2.0
pip==24.0.0
psycopg2-binary==2.9.9
pycparser==2.22
python-dotenv==1.0.1
requests==2.31.0
simpletexting==0.0.1
sqlparse==0.4.4
supabase==2.4.3
urllib3==2.2.0
waitress==3.0.0
webflow==1.2.0
